fix(ci): split where-chip tests under the lint line budget
